# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Autobot Backend API
  version: 0.1.0
  description: |
    Initial Stage 1 contract shared by the grammY bot and standalone Web Admin.
    Browser calls use the session cookie. Bot calls use the HMAC headers defined
    in ADR 0003 and include X-Telegram-User-Id for user-scoped operations.
    GPT conversation transcripts and generated image content are not persisted.
servers:
  - url: https://api.example.invalid/api/v1
    description: Public Web Admin boundary; configured per environment
  - url: http://autobot-api:3000/api/v1
    description: Internal Docker network boundary for autobot-bot
...
security:
  - cookieSession: []
  - botSignature: []
paths:
  /auth/telegram/start:
    get:
      tags: [Auth]
      operationId: startTelegramLogin
      summary: Start Telegram OIDC Authorization Code + PKCE login
      security: []
      parameters:
        - name: returnTo
          in: query
          required: true
          schema:
            type: string
            format: uri
      responses:
        "302":
          description: Redirect to Telegram authorization endpoint
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/telegram/callback:
    get:
      tags: [Auth]
      operationId: completeTelegramLogin
      summary: Validate Telegram callback and establish browser session
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          required: true
          schema: { type: string }
        - name: error
          in: query
          required: false
          schema: { type: string }
      responses:
        "302":
          description: Set the session cookie and redirect to the validated Web Admin return URL
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/session:
    get:
      tags: [Auth]
      operationId: getSession
      summary: Bootstrap the current browser session
      security:
        - cookieSession: []
      responses:
        "200":
          description: Active authenticated session
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/SessionView"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/csrf:
    get:
      tags: [Auth]
      operationId: getCsrfToken
      summary: Obtain a session-bound CSRF token for browser mutations
      security:
        - cookieSession: []
      responses:
        "200":
          description: CSRF token
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [token]
                    properties:
                      token: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Revoke the current browser session
      security:
        - cookieSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204": { description: Session revoked and cookie cleared }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /channels:
    get:
      tags: [Channels]
      operationId: listChannels
      summary: List owned channels
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ChannelStatus"
      responses:
        "200":
          description: Channel page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Channels]
      operationId: createChannel
      summary: Link and verify a Telegram channel
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CsrfTokenOptional"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelCreate"
      responses:
        "201":
          $ref: "#/components/responses/ChannelResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /channels/{channelId}:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
    get:
      tags: [Channels]
      operationId: getChannel
      summary: Get one owned channel
      responses:
        "200":
          $ref: "#/components/responses/ChannelResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Channels]
      operationId: updateChannel
      summary: Update owned channel metadata
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/CsrfTokenOptional"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelUpdate"
      responses:
        "200":
          $ref: "#/components/responses/ChannelResponse"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Channels]
      operationId: archiveChannel
      summary: Archive an owned channel
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/CsrfTokenOptional"
      responses:
        "204": { description: Channel archived }
        "409":
          $ref: "#/components/responses/Conflict"
  /channels/{channelId}/verify:
    post:
      tags: [Channels]
      operationId: verifyChannel
      summary: Revalidate Telegram channel access and bot permissions
      parameters:
        - $ref: "#/components/parameters/ChannelId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CsrfTokenOptional"
      responses:
        "200":
          $ref: "#/components/responses/ChannelResponse"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /topics:
    get:
      tags: [Topics]
      operationId: listTopics
      summary: List owned topics
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Topic page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Topics]
      operationId: createTopic
      summary: Create a topic
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CsrfTokenOptional"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicCreate"
      responses:
        "201":
          $ref: "#/components/responses/TopicResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /topics/{topicId}:
    parameters:
      - $ref: "#/components/parameters/TopicId"
    get:
      tags: [Topics]
      operationId: getTopic
      summary: Get one owned topic
      responses:
        "200":
          $ref: "#/components/responses/TopicResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Topics]
      operationId: updateTopic
      summary: Update an owned topic
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/CsrfTokenOptional"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicUpdate"
      responses:
        "200":
          $ref: "#/components/responses/TopicResponse"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Topics]
      operationId: archiveTopic
      summary: Archive an owned topic
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/CsrfTokenOptional"
      responses:
        "204": { description: Topic archived }
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: __Host-autobot_session
      description: Opaque browser session cookie.
    botSignature:
      type: apiKey
      in: header
      name: X-Autobot-Signature
      description: HMAC signature plus the companion ADR 0003 headers.
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: '^"[0-9]+"$'
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
    CsrfTokenOptional:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required for cookie-authenticated browser mutations; omitted for HMAC-authenticated bot calls.
      schema: { type: string }
...
    ChannelId:
      name: channelId
      in: path
      required: true
      schema: { type: string, format: uuid }
    TopicId:
      name: topicId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing, invalid, or expired authentication
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Ownership or entitlement denial
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: Idempotency, concurrency, duplicate, or state conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unprocessable:
      description: Domain validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ServiceUnavailable:
      description: A required dependency or configuration is unavailable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    ChannelResponse:
      description: Channel response
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: { $ref: "#/components/schemas/Channel" }
    TopicResponse:
      description: Topic response
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: { $ref: "#/components/schemas/Topic" }
...
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
            requestId: { type: string }
    PageInfo:
      type: object
      additionalProperties: false
      required: [nextCursor, hasMore]
      properties:
        nextCursor: { type: [string, "null"] }
        hasMore: { type: boolean }
    User:
      type: object
      additionalProperties: false
      required: [id, status, locale, timezone, telegram, createdAt]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [ACTIVE, BLOCKED] }
        locale: { type: string }
        timezone: { type: string }
        telegram: { $ref: "#/components/schemas/TelegramProfile" }
        createdAt: { type: string, format: date-time }
    TelegramProfile:
      type: object
      additionalProperties: false
      required: [telegramUserId]
      properties:
        telegramUserId: { type: string, pattern: "^[0-9]+$" }
        username: { type: [string, "null"] }
        firstName: { type: [string, "null"] }
        lastName: { type: [string, "null"] }
        photoUrl: { type: [string, "null"], format: uri }
...
    SessionView:
      type: object
      additionalProperties: false
      required: [user, expiresAt]
      properties:
        user: { $ref: "#/components/schemas/User" }
        expiresAt: { type: string, format: date-time }
    ChannelStatus:
      type: string
      enum: [PENDING, ACTIVE, INVALID, ARCHIVED]
    Channel:
      type: object
      additionalProperties: false
      required: [id, telegramChatId, title, status, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        telegramChatId: { type: string, pattern: "^-?[0-9]+$" }
        title: { type: string }
        username: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/ChannelStatus" }
        verifiedAt: { type: [string, "null"], format: date-time }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ChannelCreate:
      type: object
      additionalProperties: false
      required: [telegramChatId]
      properties:
        telegramChatId: { type: string, pattern: "^-?[0-9]+$" }
    ChannelUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title: { type: string, minLength: 1, maxLength: 255 }
    ChannelPage:
      type: object
      required: [data, page]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Channel" } }
        page: { $ref: "#/components/schemas/PageInfo" }
    TopicStatus:
      type: string
      enum: [ACTIVE, ARCHIVED]
    Topic:
      type: object
      additionalProperties: false
      required: [id, name, promptTemplate, status, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: [string, "null"] }
        promptTemplate: { type: string }
        status: { $ref: "#/components/schemas/TopicStatus" }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TopicCreate:
      type: object
      additionalProperties: false
      required: [name, promptTemplate]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        description: { type: [string, "null"], maxLength: 2000 }
        promptTemplate: { type: string, minLength: 1, maxLength: 20000 }
    TopicUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        description: { type: [string, "null"], maxLength: 2000 }
        promptTemplate: { type: string, minLength: 1, maxLength: 20000 }
    TopicPage:
      type: object
      required: [data, page]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Topic" } }
        page: { $ref: "#/components/schemas/PageInfo" }
...